scripts/requirements.txt
# Python dependencies for report scripts
supabase>=2.0.0
python-dotenv>=1.0.0
